WIP for data server
HTTP GET for data in various formats: vega, sexp, csv
With this we can define vega plots with remote data instead of sending
over the entire payload.  Remote data sets are the 'native' way to
work in Vega and seem to have some built-in functionality that is
lacking in embedded data.